+++ default.nix
@@ -3,7 +3,7 @@
   python3,
   lib,
   dontCheck ? false,
-  groups ? [ ],
+  extraGroups ? [ ],
 }:
 let
   injectBuildInputs =
@@ -16,15 +16,17 @@ let
     ) attrs;
 in
 poetry2nix.mkPoetryApplication {
-  inherit dontCheck groups;
+  inherit dontCheck;
   projectDir = ./.;
   preferWheels = true;
+  groups = [ "main" ] ++ extraGroups;
   nativeCheckInputs = with python3.pkgs; [
     pytestCheckHook
     pytest-cov-stub
   ];
   overrides = poetry2nix.overrides.withDefaults (injectBuildInputs {
     warc3-wet-clueweb09 = [ "setuptools" ];
+    dtaidistance = [ "setuptools" ];
   });
   meta = with lib; {
     description = "Customizable Case-Based Reasoning (CBR) toolkit for Python with a built-in API and CLI.";