+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [0.19.0](https://github.com/wi2trier/cbrkit/compare/v0.18.3...v0.19.0) (2024-11-08)
+
+### Features
+
+* **sim/graph:** add isomorphism-based approach ([8de9597](https://github.com/wi2trier/cbrkit/commit/8de9597ff422534d87f22fef5ab365625c6238dc))
+
+### Bug Fixes
+
+* **eval:** overhaul correctness completeness ([fc530bf](https://github.com/wi2trier/cbrkit/commit/fc530bf3ea5fb22ca699895f8e1ec201b19a1781))
+* **eval:** update correctness completeness ([04ffab0](https://github.com/wi2trier/cbrkit/commit/04ffab0d94ae534ba6be3b5412e25fbe220acc90))
+
+## [0.18.3](https://github.com/wi2trier/cbrkit/compare/v0.18.2...v0.18.3) (2024-11-07)
+
+### Bug Fixes
+
+* **sim/graph:** improve networkx integration ([ead9c71](https://github.com/wi2trier/cbrkit/commit/ead9c715f76429807de85820d9b628fa75a9cf21))
+
+## [0.18.2](https://github.com/wi2trier/cbrkit/compare/v0.18.1...v0.18.2) (2024-11-07)
+
+### Bug Fixes
+
+* **sim/graph:** add support for networkx ([#196](https://github.com/wi2trier/cbrkit/issues/196)) ([e682dcd](https://github.com/wi2trier/cbrkit/commit/e682dcd1501902ef7b99af13ceb1be90c53cfbe3))
+
 ## [0.18.1](https://github.com/wi2trier/cbrkit/compare/v0.18.0...v0.18.1) (2024-10-31)